Data access for working out the version of the schema in the backing store
We don't explicitly store the version of the schema in the database.
This patch paves the way for future work that will introduce an
`orc-schema-version` table dedicated to recording it.

`store-version` works out the version of the schema  by consulting this
`orc-schema-version` table and, if the table is not found, we just
compare the database schema do we find to what we expect version 1 to
look like.

If neither the `orc-schema-version` table nor the schema we expect is
present then we report version 0.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>