Remove register-add-item from external API
Herein begins a series of patches that rework the way that Entries and
their Items are written to a Register.

Currently, `register-add-item` must be called for each item referenced
in a subsequent call to `register-append-entry`. This is annoying for
the user and the way we keep track of the added Items is slow and
cumbersome.

This patch series will rework the way Items are tracked internally and
will no longer require the user to add Items before appending the
corresponding Entries.

This patch deliberately makes it impossible to append Entries that refer
to Items for the duration of the patch series.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>