Improve item-store-add! to handle already existing items
Now that current-items is no more we rely on the Backing Store to be
able to deduplicate items for us.

This patch improves `item-store-add!` to automatically add new items and
ignore items that the Backing Store already knows about.

`item-store-add!` now returns the item-ref that has been assigned or
retrieved for the item.

Care may be taken with the item-refs returned by `item-store-add!`
because it's not generally possible to distinguish a newly assigned
item-ref from one that was retrieved from an existing item. However, if
the Backing Store transaction subsequently rolls back, any newly
assigned item-refs will become undefined and there is no way to tell
recipients to invalidate them.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>