Introduce item-set-ref!
Items freshly returned from `make-item` may now turn up in calls other
than `register-add-item`. When we add things that they depend on to the
database we'll have to retrieve or create an item-ref for them.

`item-set-ref!` allows us to put that item-ref back into the item so
that we can remember we've seen it before.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>