Introduce `with-register-transaction`
The handle returned from `make-register` and `open-register` already
provide the user with Isolated access to a Register. This patch provides
ACID-like semantics for a series of writes to a Register.

i.e. the user can ensure that a series of changes are made to a Register
atomically and without interference from other users who are trying to
make changes simultaneously.

 + Must be run inside a `with-backing-store`.

This code has a number of problems and limitations but we try to err on
the side of data safety.

 + Change `with-backing-store` so that it can no longer be nested.
   This is required to prevent internal orc state being accidentally shared
   between more than one Backing Store. The main problem is in how we
   keep track of the db-ctx and whether to use SAVEPOINT or regular
   SQLite transactions.

 + Prevent `with-register-transaction` from being nested.
   This is because we cannot currently deal with the item-ref rollback
   and allowing nested transactions would make this problem more acute.

 + The user supplied code in `with-register-transaction`'s `thunk` will
   be run with `internal-call?` set to #t. This is because
   `with-register-transaction` is implemented with
   `define-in-transaction`. This is a known problem and is recorded in
   ticket #241.

Having the wrong value for `internal-call?` during a Register
transaction is potentially unsafe because it allows the user to set an
item's item-ref from code that runs inside `with-register-transaction`.
However, the risk is minimal because it requires explicit layer piercing
by the user. `internal-call?` is only otherwise used to manage the
Backing Store transaction nesting which the user does not have an API
for.

The two other problems pertain to internal orc state that affects the
Backing Store transaction control, database handles and Item reference
accounting. Errors here could lead to corruption of the persistent data.
However, a user could reasonably expect to be able to nest
`with-backing-store` and `with-register-transaction` without causing
problems so in those cases we have to explicitly disallow it until we
can provide guarantees that it is safe.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>